requirements.txt
wheel
torch
torchvision
continuum
PyYAML
wandb
faiss-cpu
accelerate
ipython
seaborn
